requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
python-dotenv==1.0.1
requests==2.32.3